Change `progress` column from `Float` to `Double`

+++ src/ZoomHub/API/Types/Content.hs
@@ -37,7 +37,7 @@ data Content = Content
   , contentUrl       :: ContentURI
   , contentReady     :: Bool
   , contentFailed    :: Bool
-  , contentProgress  :: Float
+  , contentProgress  :: Double
   , contentShareUrl  :: ContentShareURI
   , contentEmbedHtml :: String
   , contentDzi       :: Maybe DeepZoomImage
